relax the build dependency on Gen GPU
currently, the Gen GPU pciid of the underlying system is queried
and then passed to gbe_bin_generater as the target option.

This does not work when building the driver on another system with
non-intel GPUs, this patch relaxes the dependency by exporting the
pciid setting at CMake level, therefore, the pciid could be given
as a CMake option besides the current real time query method.

this patch also remove the redundancy code in utest/CMake by setting
PARENT_SCOPE in src/CMake.

Signed-off-by: Guo Yejun <yejun.guo@intel.com>
Reviewed-by: Zhigang Gong <zhigang.gong@linux.intel.com>