use LLVM_INSTALL_DIR as the path to clang/llvm-as/llvm-link
I invented CMAKE_BINARY_PATH as the path to clang/llvm-as/llvm-link
in last patch, it is not elegant. Actually, LLVM_INSTALL_DIR is
already used in CMake file and is a better choice.

So, for cross compile case, cmake can find the binaries such as clang,
llvm-as, llvm-link and llvm-config with the help of LLVM_INSTALL_DIR.

Signed-off-by: Guo Yejun <yejun.guo@intel.com>