fix opencv_test_imgproc subcase OCL_ImgProc/Accumulate.Mask regression.
This regression is caused by structural analysis when check the if-then
node, acturally there are four types of if-then node according to the
topology and fallthrough information. fallthrough check is added in this
patch.

v2: add inversePredicate member and function for BranchInstruction;
print the exact meanning of IF instruction in GEN_IR.

Signed-off-by: Luo Xionghu <xionghu.luo@intel.com>
Reviewed-by: Zhigang Gong <zhigang.gong@linux.intel.com>