GBE: work around error reporting for unresolved symbols
Beignet currently doesn't have a good error reporting mechanism
for internal passes. Currently, almost all error will cause an
user unfriendly assert. This is ok for most of the cases, but
it is really not good for unresolve symbols error, as it is not
a real compiler internal bug and we should not assert. We should
report it to help user to identify the error and fix that in the
cl kernel.

This patch is just a work around. We will implement a better
error handling in the future to consolidate all this type of
error into the normal error log buffer.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>