fix issue to create cl image from libva with non-zero offset
Beignet accepts buffer object name to share data between libva,
it supports to create cl image from the bo name with a non-zero
offset, but it does not work at some platforms.

The driver calls intel_bo_gem_create_from_name to retrieve the
dri_bo, and the offset of dri_bo is changed by the non-zero offset.
At some platforms, the change of the offset has side effect when
the kernel is executed again and so intel_bo_gem_create_from_name
is called for the second time.

So, do not change the offset of dri_bo, but maintain the non-zero
offset in cl_image, and maintain the non-zero offset until we write
the surface state into batch buffer.

V2: correct the offset parameter passed to dri_bo_emit_reloc
Signed-off-by: Guo Yejun <yejun.guo@intel.com>