Add unpacked ud and unpacked uw for long type.
The unpacked ud and uw is used for type conversion.
If src type and dst type are different, the hstride
in bytes must be same. So conversion for long need
to be:
MOV r1<2>:UD r2<4,4:1>:UQ  for ulong to ud
and
MOV r1<4>:UW r2<4,4:1>:UQ  for ulong to uw

Signed-off-by: Junyan He <junyan.he@linux.intel.com>
Reviewed-by: "Yang, Rong R" <rong.r.yang@intel.com>