Add functions for conversion between native and fake long.
Because on BDW, the native long's store and load need A64
instruction set. This instruction set is specially added to
support 64 bits read and write but has a big limtation, in
which the BTI parameter should refer to a stateless surface.
This is unaccepitable for us because it can cause overwrite
problems. We fallback to use the old manner which read and
write the long as the vec2 of int into/from top and bottom
halves. The pack/unpacked functions here play the role of
assembling/disassmebling the data before/after the write/read.

UnPack before writing like this:
  mov(8)      g108<1>:UD      g112<8,4,2>:UD        { align1 WE_normal 1Q };
  mov(8)      g110<1>:UD      g112.1<8,4,2>:UD      { align1 WE_normal 1Q };
  mov(8)      g109<1>:UD      g114<8,4,2>:UD        { align1 WE_normal 2Q };
  mov(8)      g111<1>:UD      g114.1<8,4,2>:UD      { align1 WE_normal 2Q };
  send(16)    null:UW         g106<8,8,1>:UD

and Pack after reading like this:
  send(16)    g120<1>:UW      g124<8,8,1>:UD
  mov(8)      g116<2>:UD      g120<4,4,1>:UD        { align1 WE_normal 1Q };
  mov(8)      g118<2>:UD      g121<4,4,1>:UD        { align1 WE_normal 1Q };
  mov(8)      g116.1<2>:UD    g122<4,4,1>:UD        { align1 WE_normal 1Q };
  mov(8)      g118.1<2>:UD    g123<4,4,1>:UD        { align1 WE_normal 1Q };

Signed-off-by: Junyan He <junyan.he@linux.intel.com>
Reviewed-by: Zhigang Gong <zhigang.gong@linux.intel.com>